requirements.txt
python-dotenv
python-telegram-bot
asyncio
requests
transmission_rpc
